+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Source operator: inner fields of an expression field are also added now.
+
 ## [0.7.0] - 2026-07-16
 
 ### Changed